+++ R/utils-plm.R
@@ -10,10 +10,6 @@ new_plm <- function(x, ..., class = character()) {
   structure(x, ..., class = c(class, "plm", "character"))
 }
 
-new_plm_agt <- function(x, name) {
-  new_plm(x, name = name, class = "plm_agt")
-}
-
 compare_proxy.plm <- function(x, path = "x") {
   x <- unstructure(x)
   NextMethod()